requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.1
supabase>=2.6.0
werkzeug>=3.0.0
itsdangerous>=2.2.0
gunicorn>=22.0.0
psycopg2-binary>=2.9.9
PyJWT>=2.8
requests>=2.25.0
